Departamento & Municipio Models

+++ lib/helpers/api_helper.dart
@@ -3,18 +3,59 @@ import 'package:http/http.dart' as http;
 import 'package:surveycat_app/models/parcela.dart';
 
 import 'package:surveycat_app/models/response.dart';
+import '../models/catDepartamento.dart';
+import '../models/token.dart';
 import '../models/user.dart';
 import 'constants.dart';
 
 class ApiHelper {
-  static Future<Response> getParcelas(String token) async {
+  static Future<Response> getDepartamentos(Token token) async {
+    if (!validToken(token)) {
+      return Response(
+          isSuccess: false,
+          message:
+              'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
+    var url = Uri.parse('${Constants.apiUrl}/api/CatDepartamentos');
+    var response = await http.get(
+      url,
+      headers: {
+        'content-type': 'application/json',
+        'accept': 'application/json',
+        'authorization': 'bearer ${token.token}',
+      },
+    );
+
+    var body = response.body;
+    if (response.statusCode >= 400) {
+      return Response(isSuccess: false, message: body);
+    }
+
+    List<CatDepartamento> list = [];
+    var decodedJson = jsonDecode(body);
+    if (decodedJson != null) {
+      for (var item in decodedJson) {
+        list.add(CatDepartamento.fromJson(item));
+      }
+    }
+
+    return Response(isSuccess: true, result: list);
+  }
+
+  static Future<Response> getParcelas(Token token) async {
+    if (!validToken(token)) {
+      return Response(
+          isSuccess: false,
+          message:
+              'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
     var url = Uri.parse('${Constants.apiUrl}/api/Parcelas');
     var response = await http.get(
       url,
       headers: {
         'content-type': 'application/json',
         'accept': 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
     );
 
@@ -34,14 +75,20 @@ class ApiHelper {
     return Response(isSuccess: true, result: list);
   }
 
-  static Future<Response> getUsers(String token) async {
+  static Future<Response> getUsers(Token token) async {
+    if (!validToken(token)) {
+      return Response(
+          isSuccess: false,
+          message:
+              'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
     var url = Uri.parse('${Constants.apiUrl}/api/Users');
     var response = await http.get(
       url,
       headers: {
         'content-type': 'application/json',
         'accept': 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
     );
 
@@ -61,15 +108,48 @@ class ApiHelper {
     return Response(isSuccess: true, result: list);
   }
 
+  static Future<Response> getUser(Token token, String id) async {
+    if (!validToken(token)) {
+      return Response(
+          isSuccess: false,
+          message:
+              'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
+    var url = Uri.parse('${Constants.apiUrl}/api/Users/$id');
+    var response = await http.get(
+      url,
+      headers: {
+        'content-type': 'application/json',
+        'accept': 'application/json',
+        'authorization': 'bearer ${token.token}',
+      },
+    );
+
+    var body = response.body;
+    if (response.statusCode >= 400) {
+      return Response(isSuccess: false, message: body);
+    }
+
+    var decodedJson = jsonDecode(body);
+    return Response(isSuccess: true, result: User.fromJson(decodedJson));
+  }
+
   static Future<Response> put(String controller, String id,
-      Map<String, dynamic> request, String token) async {
+      Map<String, dynamic> request, Token token) async {
+    if (!validToken(token)) {
+      return Response(
+          isSuccess: false,
+          message:
+              'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
+
     var url = Uri.parse('${Constants.apiUrl}$controller$id');
     var response = await http.put(
       url,
       headers: {
         'content-type': 'application/json',
         'accept': 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
       body: jsonEncode(request),
     );
@@ -82,14 +162,20 @@ class ApiHelper {
   }
 
   static Future<Response> post(
-      String controller, Map<String, dynamic> request, String token) async {
+      String controller, Map<String, dynamic> request, Token token) async {
+    if (!validToken(token)) {
+      return Response(
+          isSuccess: false,
+          message:
+              'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
     var url = Uri.parse('${Constants.apiUrl}$controller');
     var response = await http.post(
       url,
       headers: {
         'content-type': 'application/json',
         'accept': 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
       body: jsonEncode(request),
     );
@@ -102,14 +188,20 @@ class ApiHelper {
   }
 
   static Future<Response> delete(
-      String controller, String id, String token) async {
+      String controller, String id, Token token) async {
+    if (!validToken(token)) {
+      return Response(
+          isSuccess: false,
+          message:
+              'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
     var url = Uri.parse('${Constants.apiUrl}$controller$id');
     var response = await http.delete(
       url,
       headers: {
         'content-type': 'application/json',
         'accept': 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
     );
 
@@ -118,5 +210,12 @@ class ApiHelper {
     }
 
     return Response(isSuccess: true);
+  }
+
+  static bool validToken(Token token) {
+    if (DateTime.parse(token.expiration).isAfter(DateTime.now())) {
+      return true;
+    }
+    return false;
   }
 }
